fix(billing): enforce live allowances and meter text generation fractionally

+++ services/url_consumer_launch.py
@@ -36,7 +36,7 @@ async def launch_quoted_url_audit(*, merchant_id, product_keys, idempotency_key,
             raise UrlLaunchUnavailable('Debit exists without its matching audit; use a new quote later')
         options = deepcopy(request_options_jsonb)
         options['launch']['debited'] = [{'kind': 'audit', 'amount': credits, 'replay': False,
-            'purchased_credits': int((result.get('debit') or {}).get('purchased_credits_debited') or 0)}]
+            'purchased_credits': float((result.get('debit') or {}).get('purchased_credits_debited') or 0)}]
         run_id, replay = await enqueue_audit_run_with_replay(
             merchant_id=merchant_id, product_keys=product_keys, subject_type='merchant_url',
             idempotency_key=idempotency_key, request_options_jsonb=options)
